Fix refactoring errors
